Time: 78 ms (80.67%), Space: 42.8 MB (6.52%) - LeetHub

+++ 1508-range-sum-of-sorted-subarray-sums/1508-range-sum-of-sorted-subarray-sums.cpp
@@ -19,7 +19,7 @@ public:
             }
         }
         sort(v.begin(), v.end());
-        long long res = 0;
+        int res = 0;
        
         for(int i = left-1;i<right;i++){
             res = (res + v[i]) % mod;
